requirements.txt
docker
pytest
requests
jinja2
pyyaml
